Don't save children property to disk
Serializing items is easier when children are not written to disk,
because they create the need to write three items at once. E.g. when a
profile's distro attribute is switched from old_distro to new_distro,
old_distro and new_distro need to be saved to update the "children" json
key. Without serializing this attribute, only the profile needs to be changed.

Note: This implementation does not cache the computed list of children.
That's probably a bad idea.